backend/pyproject.toml
[project]
name = "ishuttle-backend"
version = "0.1.0"
description = "iShuttle — 배드민턴 복식 AI 분석 백엔드"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "ultralytics>=8.2.0",
    "mediapipe>=0.10.0",
    "anthropic>=0.34.0",
    "moviepy>=1.0.3",
    "ffmpeg-python>=0.2.0",
    "opencv-python>=4.10.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "torch>=2.3.0",
    "websockets>=13.0",
    "pillow>=11.3.0",
    "lapx>=0.9.4",
    "parse>=1.21.1",
    "tqdm>=4.67.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipeline", "api"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
